When installations are in progress only delay cleaning up of Temp directory, don't delay cleaning up of old extension versions.
This makes the effect of chrome never finishing installing pending extension less severe since most data will still be cleaned up.

BUG=329943

Review URL: https://codereview.chromium.org/131153003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@244720 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: e8107c161edcaea4bc628564dee39e6aa0bfe01b

+++ chrome/common/extensions/extension_file_util.cc
@@ -325,7 +325,8 @@ std::set<base::FilePath> GetBrowserImagePaths(const Extension* extension) {
 
 void GarbageCollectExtensions(
     const base::FilePath& install_directory,
-    const std::multimap<std::string, base::FilePath>& extension_paths) {
+    const std::multimap<std::string, base::FilePath>& extension_paths,
+    bool clean_temp_dir) {
   // Nothing to clean up if it doesn't exist.
   if (!base::DirectoryExists(install_directory))
     return;
@@ -343,7 +344,8 @@ void GarbageCollectExtensions(
     // Clean up temporary files left if Chrome crashed or quit in the middle
     // of an extension install.
     if (basename.value() == kTempDirectoryName) {
-      base::DeleteFile(extension_path, true);  // Recursive
+      if (clean_temp_dir)
+        base::DeleteFile(extension_path, true);  // Recursive
       continue;
     }
